+++ common/replication/deliver.go
@@ -4,7 +4,7 @@ Copyright IBM Corp. 2018 All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
 */
 
-package cluster
+package replication
 
 import (
 	"context"
