+++ common/replication/mocks/signer_serializer.go
