+++ common/replication/util.go
@@ -1,0 +1,183 @@
+/*
+Copyright IBM Corp. 2017 All Rights Reserved.
+
+SPDX-License-Identifier: Apache-2.0
+*/
+
+package replication
+
+import (
+	"crypto/x509"
+	"encoding/json"
+	"encoding/pem"
+	"fmt"
+	"time"
+
+	"github.com/hyperledger/fabric-protos-go/common"
+	"github.com/hyperledger/fabric/bccsp"
+	"github.com/hyperledger/fabric/common/channelconfig"
+	"github.com/hyperledger/fabric/internal/pkg/comm"
+	"github.com/hyperledger/fabric/protoutil"
+	"github.com/pkg/errors"
+	"google.golang.org/grpc"
+)
+
+// StandardDialer wraps an ClientConfig, and provides
+// a means to connect according to given EndpointCriteria.
+type StandardDialer struct {
+	Config comm.ClientConfig
+}
+
+// Dial dials an address according to the given EndpointCriteria
+func (dialer *StandardDialer) Dial(endpointCriteria EndpointCriteria) (*grpc.ClientConn, error) {
+	clientConfigCopy := dialer.Config
+	clientConfigCopy.SecOpts.ServerRootCAs = endpointCriteria.TLSRootCAs
+
+	return clientConfigCopy.Dial(endpointCriteria.Endpoint)
+}
+
+// BlockSequenceVerifier verifies that the given consecutive sequence
+// of blocks is valid.
+type BlockSequenceVerifier func(blocks []*common.Block, channel string) error
+
+// Dialer creates a gRPC connection to a remote address
+type Dialer interface {
+	Dial(endpointCriteria EndpointCriteria) (*grpc.ClientConn, error)
+}
+
+// EndpointCriteria defines criteria of how to connect to a remote orderer node.
+type EndpointCriteria struct {
+	Endpoint   string   // Endpoint of the form host:port
+	TLSRootCAs [][]byte // PEM encoded TLS root CA certificates
+}
+
+// String returns a string representation of this EndpointCriteria
+func (ep EndpointCriteria) String() string {
+	var formattedCAs []interface{}
+	for _, rawCAFile := range ep.TLSRootCAs {
+		var bl *pem.Block
+		pemContent := rawCAFile
+		for {
+			bl, pemContent = pem.Decode(pemContent)
+			if bl == nil {
+				break
+			}
+			cert, err := x509.ParseCertificate(bl.Bytes)
+			if err != nil {
+				break
+			}
+
+			issuedBy := cert.Issuer.String()
+			if cert.Issuer.String() == cert.Subject.String() {
+				issuedBy = "self"
+			}
+
+			info := make(map[string]interface{})
+			info["Expired"] = time.Now().After(cert.NotAfter)
+			info["Subject"] = cert.Subject.String()
+			info["Issuer"] = issuedBy
+			formattedCAs = append(formattedCAs, info)
+		}
+	}
+
+	formattedEndpointCriteria := make(map[string]interface{})
+	formattedEndpointCriteria["Endpoint"] = ep.Endpoint
+	formattedEndpointCriteria["CAs"] = formattedCAs
+
+	rawJSON, err := json.Marshal(formattedEndpointCriteria)
+	if err != nil {
+		return fmt.Sprintf("{\"Endpoint\": \"%s\"}", ep.Endpoint)
+	}
+
+	return string(rawJSON)
+}
+
+// EndpointconfigFromConfigBlock retrieves TLS CA certificates and endpoints
+// from a config block.
+func EndpointconfigFromConfigBlock(block *common.Block, bccsp bccsp.BCCSP) ([]EndpointCriteria, error) {
+	if block == nil {
+		return nil, errors.New("nil block")
+	}
+	envelopeConfig, err := protoutil.ExtractEnvelope(block, 0)
+	if err != nil {
+		return nil, err
+	}
+
+	bundle, err := channelconfig.NewBundleFromEnvelope(envelopeConfig, bccsp)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed extracting bundle from envelope")
+	}
+	msps, err := bundle.MSPManager().GetMSPs()
+	if err != nil {
+		return nil, errors.Wrap(err, "failed obtaining MSPs from MSPManager")
+	}
+	ordererConfig, ok := bundle.OrdererConfig()
+	if !ok {
+		return nil, errors.New("failed obtaining orderer config from bundle")
+	}
+
+	mspIDsToCACerts := make(map[string][][]byte)
+	var aggregatedTLSCerts [][]byte
+	for _, org := range ordererConfig.Organizations() {
+		// Validate that every orderer org has a corresponding MSP instance in the MSP Manager.
+		msp, exists := msps[org.MSPID()]
+		if !exists {
+			return nil, errors.Errorf("no MSP found for MSP with ID of %s", org.MSPID())
+		}
+
+		// Build a per org mapping of the TLS CA certs for this org,
+		// and aggregate all TLS CA certs into aggregatedTLSCerts to be used later on.
+		var caCerts [][]byte
+		caCerts = append(caCerts, msp.GetTLSIntermediateCerts()...)
+		caCerts = append(caCerts, msp.GetTLSRootCerts()...)
+		mspIDsToCACerts[org.MSPID()] = caCerts
+		aggregatedTLSCerts = append(aggregatedTLSCerts, caCerts...)
+	}
+
+	endpointsPerOrg := perOrgEndpoints(ordererConfig, mspIDsToCACerts)
+	if len(endpointsPerOrg) > 0 {
+		return endpointsPerOrg, nil
+	}
+
+	return globalEndpointsFromConfig(aggregatedTLSCerts, bundle), nil
+}
+
+func perOrgEndpoints(ordererConfig channelconfig.Orderer, mspIDsToCerts map[string][][]byte) []EndpointCriteria {
+	var endpointsPerOrg []EndpointCriteria
+
+	for _, org := range ordererConfig.Organizations() {
+		for _, endpoint := range org.Endpoints() {
+			endpointsPerOrg = append(endpointsPerOrg, EndpointCriteria{
+				TLSRootCAs: mspIDsToCerts[org.MSPID()],
+				Endpoint:   endpoint,
+			})
+		}
+	}
+
+	return endpointsPerOrg
+}
+
+func globalEndpointsFromConfig(aggregatedTLSCerts [][]byte, bundle *channelconfig.Bundle) []EndpointCriteria {
+	var globalEndpoints []EndpointCriteria
+	for _, endpoint := range bundle.ChannelConfig().OrdererAddresses() {
+		globalEndpoints = append(globalEndpoints, EndpointCriteria{
+			Endpoint:   endpoint,
+			TLSRootCAs: aggregatedTLSCerts,
+		})
+	}
+	return globalEndpoints
+}
+
+// ErrSkipped denotes that replicating a chain was skipped
+var ErrSkipped = errors.New("skipped")
+
+// ErrForbidden denotes that an ordering node refuses sending blocks due to access control.
+var ErrForbidden = errors.New("forbidden pulling the channel")
+
+// ErrServiceUnavailable denotes that an ordering node is not servicing at the moment.
+var ErrServiceUnavailable = errors.New("service unavailable")
+
+// ErrNotInChannel denotes that an ordering node is not in the channel
+var ErrNotInChannel = errors.New("not in the channel")
+
+var ErrRetryCountExhausted = errors.New("retry attempts exhausted")
